+++ src/Chorus/UI/Clone/GetCloneFromUsbDialog.Designer.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using SIL.Progress;
 using SIL.Windows.Forms.Progress;
 
@@ -48,7 +48,7 @@ namespace Chorus.UI.Clone
 			this._statusLabel = new System.Windows.Forms.TextBox();
 			this._logBox = new SIL.Windows.Forms.Progress.LogBox();
 			this._helpProvider = new Vulcan.Uczniowie.HelpProvider.HelpComponent(this.components);
-			this.l10NSharpExtender1 = new L10NSharp.UI.L10NSharpExtender(this.components);
+			this.l10NSharpExtender1 = new L10NSharp.Windows.Forms.L10NSharpExtender(this.components);
 			((System.ComponentModel.ISupportInitialize)(this.l10NSharpExtender1)).BeginInit();
 			this.SuspendLayout();
 			//
@@ -260,7 +260,7 @@ namespace Chorus.UI.Clone
 		private System.Windows.Forms.TextBox _statusLabel;
 		private LogBox _logBox;
 		private Vulcan.Uczniowie.HelpProvider.HelpComponent _helpProvider;
-		private L10NSharp.UI.L10NSharpExtender l10NSharpExtender1;
+		private L10NSharp.Windows.Forms.L10NSharpExtender l10NSharpExtender1;
 
 	}
 }